day_02_image_classification_api/requirements.txt
# Core ML & Vision
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Demo UI
streamlit>=1.28.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
httpx>=0.24.0

# Utilities
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0
